+++ demo/paging/paging-virtual.component.ts
@@ -5,16 +5,16 @@ import {CorporateEmployee} from "./model/corporate-employee";
 import {Page} from "./model/page";
 
 @Component({
-  selector: 'server-paging-demo',
+  selector: 'virtual-paging-demo',
   providers: [
       MockServerResultsService
   ],
   template: `
     <div>
       <h3>
-        Server-side Paging
+        Virtual Server-side Paging
         <small>
-          <a href="https://github.com/swimlane/ngx-datatable/blob/master/demo/paging/paging-server.component.ts" target="_blank">
+          <a href="https://github.com/swimlane/ngx-datatable/blob/master/demo/paging/paging-virtual.component.ts" target="_blank">
             Source
           </a>
         </small>
@@ -25,40 +25,55 @@ import {Page} from "./model/page";
         [columns]="[{name:'Name'},{name:'Gender'},{name:'Company'}]"
         [columnMode]="'force'"
         [headerHeight]="50"
+        [scrollbarV]="true"
         [footerHeight]="50"
-        [rowHeight]="'auto'"
+        [rowHeight]="50"
         [externalPaging]="true"
         [count]="page.totalElements"
         [offset]="page.pageNumber"
-        [limit]="page.size"
         (page)='setPage($event)'>
       </ngx-datatable>
     </div>
   `
 })
-export class ServerPagingComponent {
+export class VirualPagingComponent {
 
   page = new Page();
   rows = new Array<CorporateEmployee>();
+  cache: any = {};
 
   constructor(private serverResultsService: MockServerResultsService) {
     this.page.pageNumber = 0;
-    this.page.size = 20;
-  }
-
-  ngOnInit() {
-    this.setPage({ offset: 0 });
   }
 
   /**
    * Populate the table with new data based on the page number
    * @param page The page to select
    */
-  setPage(pageInfo){
+  setPage(pageInfo) {
     this.page.pageNumber = pageInfo.offset;
+    this.page.size = pageInfo.pageSize;
+
+    // cache results
+    // if(this.cache[this.page.pageNumber]) return;
+
     this.serverResultsService.getResults(this.page).subscribe(pagedData => {
       this.page = pagedData.page;
-      this.rows = pagedData.data;
+
+      // calc start
+      const start = this.page.pageNumber * this.page.size;
+
+      // copy rows
+      const rows = [...this.rows];
+
+      // insert rows into new position
+      rows.splice(start, 0, ...pagedData.data);
+
+      // set rows to our new rows
+      this.rows = rows;
+
+      // add flag for results
+      this.cache[this.page.pageNumber] = true;
     });
   }