+++ src/components/datatable.component.ts
@@ -20,7 +20,7 @@ import { DatatableFooterDirective } from './footer';
   selector: 'ngx-datatable',
   template: `
     <div
-      visibility-observer
+      visibilityObserver
       (visible)="recalculate()">
       <datatable-header
         *ngIf="headerHeight"
@@ -46,6 +46,7 @@ import { DatatableFooterDirective } from './footer';
         [scrollbarV]="scrollbarV"
         [scrollbarH]="scrollbarH"
         [loadingIndicator]="loadingIndicator"
+        [externalPaging]="externalPaging"
         [rowHeight]="rowHeight"
         [rowCount]="rowCount"
         [offset]="offset"
@@ -716,7 +717,19 @@ export class DatatableComponent implements OnInit, AfterViewInit, DoCheck {
 
     // this has to be done to prevent the change detection
     // tree from freaking out because we are readjusting
-    setTimeout(() => this.recalculate());
+    setTimeout(() => {
+      this.recalculate();
+
+      // emit page for virtual server-side kickoff
+      if(this.externalPaging && this.scrollbarV) {
+        this.page.emit({
+          count: this.count,
+          pageSize: this.pageSize,
+          limit: this.limit,
+          offset: 0
+        });
+      }
+    });
   }
 
   /**
@@ -770,9 +783,10 @@ export class DatatableComponent implements OnInit, AfterViewInit, DoCheck {
    *
    * @memberOf DatatableComponent
    */
-  recalculateColumns(columns: any[] = this.columns,
-                     forceIdx: number = -1,
-                     allowBleed: boolean = this.scrollbarH): any[] {
+  recalculateColumns(
+    columns: any[] = this.columns,
+    forceIdx: number = -1,
+    allowBleed: boolean = this.scrollbarH): any[] {
 
     if (!columns) return;